requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.10.0
numpy<2.0.0,>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# OpenAI and Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0

# Vector Database
pinecone-client>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.4.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
slowapi>=0.1.7
rouge>=1.0.1

# Data Processing
pyarrow>=12.0.0
jsonlines>=3.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
wordcloud>=1.9.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.0.0
loguru>=0.7.0
rich>=13.0.0

# HTTP and Utilities
httpx>=0.24.0
requests>=2.28.0
tqdm>=4.64.0
nltk>=3.8.1
tenacity>=8.2.0
aiohttp>=3.8.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Deployment
gunicorn>=21.0.0 